Release automation on GitHub (#8)
this commit is a .github/workflows/gradle.yml file. This will cause github to build the code on commit, and if you push a git tag to git, it will do a release based on that tag. If the commit is not tagged, the build will still run and the final .zip file will be available as an artifact, found under the "actions" button in github. (#8)